Allow fractional ENGINE_TICK_SECONDS instead of crashing on one
Was typed int, so any non-integer value (someone trying "0.5", say) would
fail Pydantic validation and crash the Celery worker on startup rather
than just tick unexpectedly. Now float, with a comment on why going much
below ~2s is counterproductive rather than just "faster": each tick is a
real HTTP round trip per running bot to whatever broker connection it
uses, and sub-2s scheduling risks ticks queuing up faster than they
complete instead of actually syncing faster.

+++ tilly-trading/backend/app/core/config.py
@@ -62,7 +62,11 @@ class Settings(BaseSettings):
 
     # ----- Bot engine -----
     # Seconds between engine ticks (dispatch + on_tick) in the Celery beat loop.
-    engine_tick_seconds: int = 5
+    # float, not int: each tick makes a real HTTP round trip per running bot to
+    # whatever broker connection it uses (for the MT5 bridge: Render -> tunnel
+    # -> the user's PC -> MT5 IPC -> back), so going much below ~2s risks ticks
+    # queuing up faster than they complete rather than actually syncing faster.
+    engine_tick_seconds: float = 5.0
 
     # ----- AI presets (optional) -----
     openai_api_key: str = ""
